pyproject.toml
[project]
name = "recon-ng-mcp-server"
version = "0.1.0"
description = "Model Context Protocol server for recon-ng OSINT framework"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.12.4",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
recon-mcp = "recon_ng_mcp_server.mcp_server:mcp"
recon-mcp-run = "recon_ng_mcp_server.mcp_server:main"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
